+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -2,6 +2,7 @@ export 'facebook.dart';
 export 'instagram.dart';
 export 'linkedin.dart';
 export 'pinterest.dart';
+export 'slack.dart';
 export 'telegram.dart';
 export 'tiktok.dart';
 export 'twitter.dart';